Moved TC_SUTestBase to support module. References updated.
