+++ src/python_testing/TC_SU_4_1.py
@@ -41,7 +41,7 @@
 import logging
 
 from mobly import asserts
-from TC_SUTestBase import SoftwareUpdateBaseTest
+from support_modules.su_support import SoftwareUpdateBaseTest
 
 import matter.clusters as Clusters
 from matter import ChipDeviceCtrl
